refactor: 관심사에 맞게 패키지 이동
webserver: Socket, Http, 멀티 스레드 요청 위임과 관련된 기능
servlet: 요청에 대한 실제 처리 로직 기능
