+++ src/main/java/codesquad/Main.java
@@ -1,7 +1,7 @@
 package codesquad;
 
+import codesquad.configuration.Container;
 import codesquad.webserver.Connector;
-import codesquad.webserver.Container;
 
 public class Main {
     public static void main(String[] args) {
