+++ src/main/java/codesquad/servlet/handler/resource/SupportFileExtension.java
@@ -1,4 +1,4 @@
-package codesquad.webserver;
+package codesquad.servlet.handler.resource;
 
 public enum SupportFileExtension {
