+++ src/main/java/codesquad/webserver/parser/HttpRequestMapper.java
@@ -1,5 +1,9 @@
-package codesquad.webserver.http;
+package codesquad.webserver.parser;
 
+import codesquad.webserver.http.HttpHeaders;
+import codesquad.webserver.http.HttpRequest;
+import codesquad.webserver.http.HttpRequestBody;
+import codesquad.webserver.http.HttpRequestLine;
 import java.io.BufferedReader;
 import java.io.IOException;
 import org.slf4j.Logger;
