+++ src/main/java/codesquad/webserver/parser/HttpRequestParser.java
@@ -1,5 +1,11 @@
-package codesquad.webserver.http;
+package codesquad.webserver.parser;
 
+import codesquad.webserver.http.HttpHeaders;
+import codesquad.webserver.http.HttpMethod;
+import codesquad.webserver.http.HttpPath;
+import codesquad.webserver.http.HttpRequestBody;
+import codesquad.webserver.http.HttpRequestLine;
+import codesquad.webserver.http.HttpVersion;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.net.URLDecoder;
